requirements.txt
# Core Data Science Libraries
pandas==2.1.4
numpy==1.26.2
scipy==1.11.4

# Machine Learning
scikit-learn==1.3.2
implicit==0.7.2

# Configuration
pyyaml==6.0.1

# Data Versioning
dvc==3.39.0
dvc-s3==3.2.2

# API Framework
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3

# Caching
redis==5.0.1

# Experiment Tracking
mlflow==2.9.2

# Monitoring
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Code Quality
black==23.12.1
flake8==7.0.0

# Visualization (for Streamlit)
streamlit==1.30.0
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.1

# Utilities
python-dotenv==1.0.0
requests==2.31.0
